refactor(mcp): split MCP into Uri/Handlers/Server submodules
Emanote.MCP becomes an umbrella module re-exporting 'run'. Work
lives in submodules decomposed by concern:

- Emanote.MCP.Uri — wire-contract URI constants + ResourceKind↔URI
  translation (pure, no MCP type deps)
- Emanote.MCP.Handlers — request handlers + Catalog→MCP wire-type
  adapters (toMcpResource, textResult, noteTemplate)
- Emanote.MCP.Server — Warp startup + server identity / capabilities
  / instructions

Zero behavior change at the MCP wire.
