+++ emanote/emanote.cabal
@@ -166,6 +166,9 @@ library
     Emanote.CLI
     Emanote.MCP
     Emanote.MCP.Catalog
+    Emanote.MCP.Handlers
+    Emanote.MCP.Server
+    Emanote.MCP.Uri
     Emanote.Model
     Emanote.Model.Calendar
     Emanote.Model.Calendar.Parser
